+++ lib/firebase_github_updater.dart
@@ -3,3 +3,4 @@ export 'services/download_service.dart';
 export 'services/install_service.dart';
 export 'models/app_release.dart';
 export 'widgets/update_bottom_sheet.dart';
+export 'firebase_github_updater_helper.dart';